first complete pass review
