+++ config/components/manager/controller_manager_config.yaml
@@ -28,9 +28,6 @@ clientConnection:
 #  enable: false
 #  webhookServiceName: ""
 #  webhookSecretName: ""
-queueVisibility:
-  clusterQueues:
-    maxCount: 10
 integrations:
   frameworks:
   - "batch/job"
