+++ pkg/config/config_test.go
@@ -18,7 +18,6 @@ package config
 
 import (
 	"errors"
-	"fmt"
 	"io/fs"
 	"os"
 	"path/filepath"
@@ -29,7 +28,6 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/apimachinery/pkg/util/yaml"
 	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -231,7 +229,7 @@ kind: Configuration
 queueVisibility:
   updateIntervalSeconds: 10
   clusterQueues:
-    maxCount: 10
+    maxCount: 0
 `), os.FileMode(0600)); err != nil {
 		t.Fatal(err)
 	}
@@ -277,6 +275,9 @@ queueVisibility:
 
 	defaultQueueVisibility := &configapi.QueueVisibility{
 		UpdateIntervalSeconds: configapi.DefaultQueueVisibilityUpdateIntervalSeconds,
+		ClusterQueues: &configapi.ClusterQueueVisibility{
+			MaxCount: 10,
+		},
 	}
 
 	testcases := []struct {
@@ -569,7 +570,7 @@ queueVisibility:
 				QueueVisibility: &configapi.QueueVisibility{
 					UpdateIntervalSeconds: 10,
 					ClusterQueues: &configapi.ClusterQueueVisibility{
-						MaxCount: 10,
+						MaxCount: 0,
 					},
 				},
 			},
@@ -669,6 +670,7 @@ func TestEncode(t *testing.T) {
 				},
 				"queueVisibility": map[string]any{
 					"updateIntervalSeconds": int64(configapi.DefaultQueueVisibilityUpdateIntervalSeconds),
+					"clusterQueues":         map[string]any{"maxCount": int64(10)},
 				},
 			},
 		},
@@ -686,53 +688,6 @@ func TestEncode(t *testing.T) {
 			}
 			if diff := cmp.Diff(tc.wantResult, gotMap); diff != "" {
 				t.Errorf("Unexpected result (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestValidate(t *testing.T) {
-	testcases := []struct {
-		name    string
-		cfg     *configapi.Configuration
-		wantErr field.ErrorList
-	}{
-
-		{
-			name:    "empty",
-			cfg:     &configapi.Configuration{},
-			wantErr: nil,
-		},
-		{
-			name: "invalid queue visibility UpdateIntervalSeconds",
-			cfg: &configapi.Configuration{
-				QueueVisibility: &configapi.QueueVisibility{
-					UpdateIntervalSeconds: 0,
-				},
-			},
-			wantErr: field.ErrorList{
-				field.Invalid(field.NewPath("queueVisibility").Child("updateIntervalSeconds"), 0, fmt.Sprintf("greater than or equal to %d", queueVisibilityClusterQueuesUpdateIntervalSeconds)),
-			},
-		},
-		{
-			name: "invalid queue visibility cluster queue max count",
-			cfg: &configapi.Configuration{
-				QueueVisibility: &configapi.QueueVisibility{
-					ClusterQueues: &configapi.ClusterQueueVisibility{
-						MaxCount: 4001,
-					},
-					UpdateIntervalSeconds: 1,
-				},
-			},
-			wantErr: field.ErrorList{
-				field.Invalid(field.NewPath("queueVisibility").Child("clusterQueues").Child("maxCount"), 4001, fmt.Sprintf("must be less than %d", queueVisibilityClusterQueuesMaxValue)),
-			},
-		},
-	}
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			if diff := cmp.Diff(tc.wantErr, validate(tc.cfg), cmpopts.IgnoreFields(field.Error{}, "BadValue")); diff != "" {
-				t.Errorf("Unexpected returned error (-want,+got):\n%s", diff)
 			}
 		})
 	}
